Found possible cause of intermittent failures
The order in which partitions are merged in `treeReduce` is not the same between different runs. Depending on
the order in which they are merged `findMates` will sometimes not find all faces that are shared. Presently, it
is unclear why they are not being detected since shared faces are detected when doing an additional check immediately
after for debugging.
Code-wise, not much is changed in this commit. Mostly print statements throughout the code to help locate cause of
failure.