Fixed NumberFormatException in `centroid`.
When volume of block is less than EPSILON, centroid gets set equal to the average value of all the
vertices that define the block. Since volume is so small, block is practically a point. Tested with
extremely tiny input cube as well as smaller Spaulding data set.
Still need to check on where all distances are getting set to EPSILON in the case where EPSILON = 1e-6.