Modified InputProcessor to accept new format of input file
Load balancing will now instead be done with real joints rather than using
artificial joints that later need to be removed. This will avoid the issues
when having to deal with treeReduce or groupBy. By not using these functions,
data shuffling is avoided.
Joints will now instead be generated with SparkRocks based on user inputs. By doing
this within SparkRocks, it will be much easier to pick appropriate joints to maintain
a somewhat balanced load among processes. The logic to do this will be placed in
JointGenerator.